pyproject.toml
[tool.poetry]
name = "book"
version = "0.1.0"
description = ""
authors = ["internet_trash <naiknischal92@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uagents = { version = "^0.11.0", python = ">=3.9,<3.13" }
twilio = "^9.0.4"
fastapi = "^0.110.1"
uvicorn = "^0.20.0"
pymongo = "^4.6.3"
python-dotenv = "^1.0.1"
pydantic = "^1.10.15"
datetime = "^5.5"
flask = {extras = ["async"], version = "^3.0.3"}



[tool.poetry.group.dev.dependencies]
requests = "^2.31.0"
mongomock = "^4.1.2"
pytest = "^8.1.1"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
